+++ src/MachinaGrasshopper/Bridge/Send.cs
@@ -79,16 +79,17 @@ namespace MachinaGrasshopper.Bridge
 
                 foreach (Machina.Action a in actions)
                 {
-                    // If attaching a tool, send the tool description first.
-                    // This is quick and dirty, a result of this component not taking the robot object as an input.
-                    // How coud this be improved...? Should tool creation be an action?
-                    if (a.type == Machina.ActionType.Attach)
-                    {
-                        ActionAttach aa = (ActionAttach)a;
-                        ins = aa.tool.ToInstruction();
-                        instructions.Add(ins);
-                        ms.socket.Send(ins);
-                    }
+                    // "DefineTool" is now an Action on its own, so shouldn't need this!
+                    //// If attaching a tool, send the tool description first.
+                    //// This is quick and dirty, a result of this component not taking the robot object as an input.
+                    //// How coud this be improved...? Should tool creation be an action?
+                    //if (a.type == Machina.ActionType.Attach)
+                    //{
+                    //    ActionAttach aa = (ActionAttach)a;
+                    //    ins = aa.tool.ToInstruction();
+                    //    instructions.Add(ins);
+                    //    ms.socket.Send(ins);
+                    //}
 
                     ins = a.ToInstruction();
                     instructions.Add(ins);